% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{A1_ADT_normal}
\alias{A1_ADT_normal}
\title{Normalized ADT expression matrix}
\format{
data frame
}
\usage{
A1_ADT_normal
}
\description{
Normalized ADT expression matrix
}
\keyword{datasets}
