% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{A1_gene_pca}
\alias{A1_gene_pca}
\title{PCA dimensionality reduction results of gene expression}
\format{
data frame
}
\usage{
A1_gene_pca
}
\description{
PCA dimensionality reduction results of gene expression
}
\keyword{datasets}
