% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{A1_true_label}
\alias{A1_true_label}
\title{Ground-truth spot cluster labels}
\format{
data frame
}
\usage{
A1_true_label
}
\description{
Ground-truth spot cluster labels
}
\keyword{datasets}
