% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSMIND.R
\name{BSMIND}
\alias{BSMIND}
\title{BSMIN function}
\usage{
BSMIND(
  data_1,
  data_2,
  coord,
  platform = c("ST", "Visium"),
  K_fixed = 3,
  a_mu = 0,
  b_mu = 1,
  IGkappa = 2,
  IGtau = 10,
  a_mu2 = 2,
  b_mu2 = 1,
  IGkappa2 = 2,
  IGtau2 = 10,
  a_beta = 0.7,
  tau_beta = 1,
  dpAlpha = 1,
  minPsi1 = 0.8,
  maxPsi1 = 0.95,
  minPsi2 = 0.8,
  maxPsi2 = 0.95,
  tau0 = 2,
  tau1 = 0.05,
  numOfMCMC = 1000,
  burnIn = 500,
  Is_print = TRUE,
  print_gap = 10,
  Is_random_seed = F,
  random_seed = 2234
)
}
\arguments{
\item{data_1}{First n*p1 data matrix (spots x genes), gene expression modality.}

\item{data_2}{Second n*p2 data matrix (spots x features), secondary feature modality.}

\item{coord}{Data frame of spot coordinates, 2 columns: x, y coordinates.}

\item{platform}{Spatial sequencing platform; "ST" (square grid) or "Visium" (hex grid).}

\item{K_fixed}{Fixed number of regions for all latent layers. Default is 3.}

\item{a_mu}{Mean of normal prior for gene expression cluster means mu_gk. Default is 0.}

\item{b_mu}{Standard deviation of normal prior for mu_gk. Default is 1.}

\item{IGkappa}{Shape parameter of inverse-gamma prior for gene expression sigma_g. Default is 2.}

\item{IGtau}{Scale parameter of inverse-gamma prior for sigma_g. Default is 10.}

\item{a_mu2}{Mean of normal prior for second modality cluster means mu_pk. Default is 2.}

\item{b_mu2}{Standard deviation of normal prior for mu_pk. Default is 1.}

\item{IGkappa2}{Shape parameter of inverse-gamma prior for second modality sigma_p. Default is 2.}

\item{IGtau2}{Scale parameter of inverse-gamma prior for sigma_p. Default is 10.}

\item{a_beta}{Prior mean for spatial Potts interaction parameter beta. Default is 0.7.}

\item{tau_beta}{Prior standard deviation for beta. Default is 1.}

\item{dpAlpha}{Dirichlet concentration parameter for cluster proportion prior pi_vec. Default is 1.}

\item{minPsi1}{Lower bound of uniform prior for gene expression contribution weight psi1. Default is 0.8.}

\item{maxPsi1}{Upper bound of uniform prior for gene expression contribution weight psi1. Default is 0.95.}

\item{minPsi2}{Lower bound of uniform prior for secondary feature modality contribution weight psi2. Default is 0.8.}

\item{maxPsi2}{Upper bound of uniform prior for secondary feature modality contribution weight psi2. Default is 0.95.}

\item{tau0}{Scale parameter for Dirichlet proposal distribution of pi_vec. Default is 2.}

\item{tau1}{Proposal standard deviation for Potts parameter beta. Default is 0.05.}

\item{numOfMCMC}{Total MCMC iterations including burn-in. Default is 5000.}

\item{burnIn}{Number of burn-in iterations (discarded for posterior inference). Default is 2500.}

\item{Is_print}{Logical; if TRUE, print iteration progress. Default is TRUE.}

\item{print_gap}{Iteration interval for progress printing. Default is 10.}

\item{Is_random_seed}{Logical; if TRUE, set fixed random seed for reproducibility. Default is FALSE.}

\item{random_seed}{Random seed number used if Is_random_seed=TRUE. Default is 2234.}
}
\value{
List containing full and post-burn-in MCMC posterior samples:
\item{mu_gk_mcmc}{Posterior samples of gene expression cluster means}
\item{sigma_g_mcmc}{Posterior samples of gene expression standard deviations}
\item{mu_pk_mcmc}{Posterior samples of second modality cluster means}
\item{sigma_p_mcmc}{Posterior samples of second modality standard deviations}
\item{clIds_mcmc}{Posterior samples of integrative cluster labels}
\item{L1_Ids_mcmc}{Posterior samples of gene-specific cluster labels}
\item{L2_Ids_mcmc}{Posterior samples of second modality-specific cluster labels}
\item{pottsBeta_mcmc}{Posterior samples of spatial interaction parameter beta}
\item{psi1_mcmc}{Posterior samples of modality 1 weight psi1}
\item{psi2_mcmc}{Posterior samples of modality 2 weight psi2}
\item{pi_mcmc}{Posterior samples of cluster proportion vector}
\item{all_mu_gk_mcmc}{List, all posterior samples including burn-in of gene expression cluster means}
\item{all_sigma_g_mcmc}{List, all posterior samples including burn-in of gene expression standard deviations}
\item{all_mu_pk_mcmc}{List, all posterior samples including burn-in of second modality cluster means}
\item{all_sigma_p_mcmc}{List, all posterior samples including burn-in of second modality standard deviations}
\item{all_clIds_mcmc}{List, all posterior samples including burn-in of integrative cluster labels}
\item{all_L1_Ids_mcmc}{List, all posterior samples including burn-in of gene-specific cluster labels}
\item{all_L2_Ids_mcmc}{List, all posterior samples including burn-in of second modality-specific cluster labels}
\item{all_pottsBeta_mcmc}{Vector, all posterior samples including burn-in of spatial interaction parameter beta}
\item{all_psi1_mcmc}{Vector, all posterior samples including burn-in of modality 1 weight psi1}
\item{all_psi2_mcmc}{Vector, all posterior samples including burn-in of modality 2 weight psi2}
\item{all_pi_mcmc}{List, all posterior samples including burn-in of cluster proportion vector}
\item{exeTime}{Total model execution time}
}
\description{
BSMIN function
}
