% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{sim1_1_data_1}
\alias{sim1_1_data_1}
\title{sim1_1 Simulated Data - Expression matrix 1}
\format{
data frame
}
\usage{
sim1_1_data_1
}
\description{
sim1_1 Simulated Data - Expression matrix 1
}
\keyword{datasets}
