% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{sim1_1_data_2}
\alias{sim1_1_data_2}
\title{Sim1_1 Simulated Data - Expression matrix 2}
\format{
data frame
}
\usage{
sim1_1_data_2
}
\description{
Sim1_1 Simulated Data - Expression matrix 2
}
\keyword{datasets}
