% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{sim1_2_data_coord}
\alias{sim1_2_data_coord}
\title{sim1_2 Simulated Data - Spatial coordinates}
\format{
data frame
}
\usage{
sim1_2_data_coord
}
\description{
sim1_2 Simulated Data - Spatial coordinates
}
\keyword{datasets}
