% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{sim2_1_data_2}
\alias{sim2_1_data_2}
\title{sim2_1 Simulated Data - Expression matrix 2}
\format{
data frame
}
\usage{
sim2_1_data_2
}
\description{
sim2_1 Simulated Data - Expression matrix 2
}
\keyword{datasets}
