% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{sim2_2_data_coord}
\alias{sim2_2_data_coord}
\title{sim2_2 Simulated Data - Spatial coordinates}
\format{
data frame
}
\usage{
sim2_2_data_coord
}
\description{
sim2_2 Simulated Data - Spatial coordinates
}
\keyword{datasets}
