% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{sim2_3_data_1}
\alias{sim2_3_data_1}
\title{sim2_3 Simulated Data - Expression matrix 1}
\format{
data frame
}
\usage{
sim2_3_data_1
}
\description{
sim2_3 Simulated Data - Expression matrix 1
}
\keyword{datasets}
