% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{sim2_3_true_label}
\alias{sim2_3_true_label}
\title{sim2_3 Simulated Data - Ground-truth labels}
\format{
data frame
}
\usage{
sim2_3_true_label
}
\description{
sim2_3 Simulated Data - Ground-truth labels
}
\keyword{datasets}
